Add files via upload
This script is the cleaned version of the manual method for the budgeting app. The first iteration was an Excel workbook showing proof of concept and some foundational design to have a category-based budget template with starting balances and a running-total remaining balance. This second iteration takes the Excel file ideas and expresses them as an R script relying on user input at every step. This serves as the basis for the Shiny app.